rework file structure + exports
