+++ src/classes/live.ts
@@ -1,7 +1,7 @@
-import Surreal from "../index.ts";
-import Emitter from "./emitter.ts";
+import { Surreal } from "../client.ts";
+import { Emitter } from "./emitter.ts";
 
-export default class Live extends Emitter<{
+export class Live extends Emitter<{
 	"create": [unknown];
 	"update": [unknown];
 	"delete": [unknown];
